Added a variable dimension to the block data so that I can run a kernel on one variable and a second kernel concurrently on a separate variable.
Initialize the block data so that the density variable is P1 and energy P2.

Coded up the runtime so that we compute the Laplacian on density and the
Laplacion on energy concurrently with two different thread teams.

Hooked up the scaling to one of the thread teams, but it is currently
incorrectly scaling both variables.  This is an error because the scaling thread
team is the work subscriber to only one of the Laplacian teams.  Therefore,
computing the Laplacian and scaling is presently a race condition.