Started adding in data packet and specifically ThreadTeams that work with data packet of block as their data type.
To make this work, I had to change the interface of the action routines so that
they are no longer templated.  Rather, they use void* and the action routines
are casting these to the type of dataItem that they are written to work with.

This was necessitated by the RuntimeAction definition.  Since these are just a
current simple test, it is not clear that the use of void* is really necessary.