(Issue #14) Added a first draft of a valgrind automated test of the thread teams.
Note that I there is a possible leak error that appears associated with
resource allocation that takes place when creating pthreads.  This
error appears to be random as sometimes there is no error and others
there could be 1 or more repeats of the error.  I have not been able to
trace the underlying cause of this and have temporarily suppressed it.