(Issue #14) WIP: Built in a logger at the level of the whole repo and adjusted code and tests.
The ThreadTeam is too central an element of thread team configurations.  I need
to update the design to accomodate thread team helpers at the same level of
ThreadTeam.  This means that the thread and data pub/sub cannot line in
ThreadTeam.  As a first step, the base, which is not a template, now contains
all thread pub/sub interface/functionality.

All tests build/run on hello.  Didn't check if the gatherData* result are valid.
However, the F2003 results seem to be scaling.