Added a class method to FArray4D so that PUDs can easily build a scratch array with the FArray4D wrapper in place.  As part of the implementation, their scratch array is allocated and deallocated automatically.
I dislike this design because the original simple, clean design concept of the
FArray4D class was that of a simple wrapper --- it was simply the lens that
allowed you to look at the memory in a different way from a distance.  Now, it
is or is not the owner of the data that you are looking at depending on how you
created it.

This complexity might be reasonable as it offers a simple mechanism to the PUDs
and shields them from the data managment.  So far, they only get access to an
FArray4D object by calling an infrastructure routine.  These routines manage the
underlying resource in such a way that, so far, the PUDs need not think about
resource management.