Reran all tests on Summit after the merge of the merge clean-up.
The slowdown originally seen in low-level performance of the ThreadTeam was not
seen this time, which is comforting given that this timing test should not have
been affected much.

Total runtimes have increased some, which is to be expected as our design now
has more layers to it.  However, the runtime speedup has increased slightly in
some cases - a pyrrhic victory.

In the C++/Fortran comparison test, I see that the runtime performance has
improved for both C++ and Fortran when only a few blocks are fed through the
pipeline.

Ran the valgrind test on GCE and did not find any leaks.