(Issue #8) Updated the data packet so that it includes loGC and hiGC. Note that these are added to the data packet as 6 different ints because we need to send over the full coordinate triple so that kernels can always be written for the most general 3D case.
Also, the IntVect class contains an array data
member.  Therefore, I cannot simply copy the IntVect to the data packet because
the copy of this in device memory would point to the array in host memory.  I
would have to copy in the IntVect object and the array.  With these in device
memory, I would have to overwrite the IntVect's array pointer so that it points
to (is attached to) the array in device memory.  Too complicated for what we
want to do and this would be too rooted in one programming model (e.g. OpenACC).