(Issue #8) At one point, I wanted to use an OpenACC runtime routine from within a CUDA file.  As part of this, I needed to inform nvcc where to find PGI's openacc.h file.  As part of this I discovered that nvcc was using the default 4.8.5 gcc Summit compiler as the backend.  In order to set pgc++ as the backend, I had to rollball the PGI version to 19.  I could have used 20 with CUDA 11, but there were issues with this as this might be too close to the bleeding edge.
Built up the installation for this PGI version with the SW stack now largely set
to the default versions of the various modules used.  This, indeed, seems like a
reasonable setup.

Note that I am also now using CUDA 10.  I realized that the -ta flags used
before were 10.1 despite the fact that I was using CUDA 9.  Note clear if this
was causing issues before.