(Issue #8)  Rolled back the code that drives data transfers and executing the kernel to a simpler state.  I wonder if the data unpacking should be done from within the kernel launch so that we don't have host code in the kernel function that could run before the asynchronous transfer finishes.  Based on the profile info, so far the transfers have always finished before the kernel starts, which means that we might be getting lucky.
To prepare for testing this hunch, I have fused the serial and loop Oacc kernels
into a single kernel.  I see in the compile message and in the profile output
only a single kernel/block.  Decorated all loops in the nest with a loop
directive as well.  The runtime/block decreased from 350 to 80 uS.