(Issue #8) The data packet now allows for copying back to UNK either the data (CC1) or scratch (CC2) data.  As a result, I could now get rid of the scratch to f_d copy back in the Density kernel.  The same can be done for the energy Laplacian kernel.  However, I didn't do it yet just to test that the old way still works.
As part of this, I had to implement a basic variable masking so that when the
data packet unpacks, it only copies back the data that was updated.  This was
necessary for the density kernel as the energy data in the scratch block was
garbage and would overwrite the energy IC without masking.  With the current
implementation, the data packet is still copying unused data to and from the
GPU.  This is, therefore, a work in progress.