(Issue #8) The fused action example as well as the stream/queue interface of the DataPacket class highlight that there is a general need for streams, regardless of the backend that will be used.  Therefore, the runtime now has generic Stream and StreamManager classes.
As part of implementing this, I decided to simplify the CudaStreamManager as
some of the paranoid internal checking is likely no longer necessary.

Updated all tests to use this and these are passing without correctness or
performance regressions.

NOTE: The F2003 gather data test is presently failing.  This will be addressed
when I try to move this test to using GPUs.