(Issue #22) First attempt at multithreading the action parallel pipeline.  Based on examples in the AMReX tutorial (v21.01) Tutorials/Amr/Advection_AmrCore/, such as AmrCoreAdv.cpp, it appears that the MFIter can be multithreaded with OpenMP simply by wrapping the creation of the MFIter and the loop that uses it with a parallel construct.  Each thread, therefore, creates its own MFIter and the partitioning of blocks must be done under the hood by the MFIter itself.
Based on the experiment in this commit, it looks like the partitioning of blocks
across the threads is done statically and such that each thread is assigned a
set of contiguous grid indices over which to iterate.

The AMReX user manual references an MFItInfo class and shows how this can be
used to do dynamic scheduling of blocks across a multithreaded MFIter loop.
There is no mention of static scheduling.  The Doxygen docs for MFItInfo don't
mention static nor what the default scheduling strategy is.