(Issue #24) WIP: Updated the runtime element so that elements calling another element's closeQueue command must now inform it of their identity so that the subscribers can now keep track of which publishers have already called closeQueue and only close the queue once all publishers have done so.
Ran the RuntimeCuda and sedovGpu 3D tests on Summit.  I had to adjust the #
blocks/packet for one setup so that we aren't using up streams so quickly (a
test was failing due to a lack of streams).  Otherwise, no regressions and
runtimes don't seem to have increased much.

There are several tests that need to be updated to use this new interface.  To
be done on a different machine.

NOTE: There are *no* tests using a TT config with a data subscriber attached to
more than one publisher.