Add two more examples for extended recipe syntax
This is an attempt to cast variants of the existing Hydro.F90 code, from the
HydroMain/unsplit implementations, into the shape of recipes.
Various additions to recipe syntax and semantics are implied.

One major complication to which these fragments demonstrate suggested
solutions is the handling of flux correction.

These recipes are based on existing logic in

    HydroMain/unsplit/Hydro.F90
    HydroMain/unsplit/levelFlux/Hydro.F90

respectively.

Not all variables that are actually used and modified by unsplit Hydro implementations
are actually mentioned here!  (EINT, TEMP, GAMC, mass scalars, and mass fractions
are some kinds of variables not mentioned.)

A way to define and name sets or groups of variables (see use of 'Vels' for VELX,VELY,VELZ)
should be developed.
A way to declare what the variabe names refer to should be developed.
It should be able to express that, for example,

   ENER means ENER_VAR in UNK,
   presStar means HY_VAR2_SCRATCHCTR_VAR in SCRATCH_CTR storage,

and so on; and similar for shapes and centerings (and rank?) of
named variables including those in SCRATCH.

Incidentally, SUSE is supposed to stand for SCRATCH USE.