Pull MPI configuration out of the runtime.  The design is that applications should determine communicators.  If a process only does work if it is the lead process, then the application should also tell it which rank is the lead rank.
Note that AMReX is still being set with MPI_COMM_WORLD, so that MPI has not yet
been moved completely out of the library.  This will be done as a different
feature.

This moved the MPI constants out of constants*.h as well, which was the main
goal.

Pulled Timer out of the runtime library since it is only used by Sedov.  Nice
simplification of the library.

Ran some tests on GCE and found that they built without error/warnings.  Ran
through as well.