Integrated a new tool into the build system that writes the library header. This means that NDIM and K[123]D could be removed from the constants*.h/Flash*.h files.
Ran tests manually on GCE and all passed.  No warnings in build logs.  Manually
inspected the setup.log and milhoja.h as well.  All looking good.

Manually tested write_library_header.py on the command line to confirm that
errors are printed correctly to stderr and that non-zero error codes are
returned.  Manually tested setup.py to confirm that error messages from the
header script are correctly written to stderr and that the setup.py script
returns a non-zero error code with the write script fails.