WIP:  First pass at establishing a good filename and namespace scheme.  Built some tests on GCE.  No new warnings in build logs.  Ran the associated tests and no obvious issues seen.  This testing has been fairly basic only and more testing will be needed.
I am generally happy with these changes.  The code is cleaner and more
homogeneous.  The delineation between library and tests is sharper.  Also, the
high-level structure and architecture of the library and its use in the
application space is clearer.

This is me paying back technical debt accumulated by kicking the can down the
road for many different low-level design decisions.  The can was kicked to avoid
unnecessary distractions from the main technical and high-level design
challenges.  Also, polishing the code and design seemed like a bad idea since we
had not yet determined that the library would be useful.  Based on this
experience, I would say that:
1) This was very tedious as one would expect.  However, going through the entire
code at a high-level has helped me assess the quality of the code as well as the
quality of my decisions.
2) Updating the library was not too difficult
3) Updating the tests was harder as they use the library a lot and the
aforementioned delineation was not sharp
4) So far the work has been fairly straightforward and low-risk
5) This would have been easier and even lower-risk if a basic filename scheme
had been adopted early on.  For example, having some files start with
Orchestration, others with Grid, and still others with no semblance of a
namespace designation in the name made the update more challenging and required
more manual work.  Having thought upfront more about the high-level structure of
the code and repo could have helped and should be done.
6) Waiting so long to pay down the technical debt was bad in the sense that
there are so many tests that had to be updated.  At the same time, waiting so
long to pay down the debt was a good idea as I have so many tests to help catch
mistakes.
7) At the moment, I believe that doing this cleaning before officially creating
the Fortran/C++ layer is a good idea.  In particular, there is less possibility
of having two files from different worlds with the same or similar names.
Ditto for variable and function names.  Part of this has been removing the bits
of the design/implementation that are too rooted in Flash or that overlap too
much with Flash.  Separating out the tests from the library helped as the
overlap was strong there.  Finally, having to clean the Fortran/C++ layer would
have made this tedious task worse and I do not foresee any benefit that could
arise from the exercise.
8) Thinking toward the Fortran/C++ layer, it will be good to rename some key
macros that are common between the library and Flash (e.g., NDIM, MDIM,
K[123]D).