The GCE CI test suite was failing on the first Sedov test, which was MPI/2D.  I confirmed that the MPI/3D and Cpu/3D were passing, but that Cpu/2D was also failing.  I found that I had neglected to update macros used in the "runtime parameter" files.  This is allowed as these files are headers that are compiled directly.  The errors in the par files is consistent with the observed pattern of failures.  Able to run the two failing tests and found reasonable integrated quantities (they were sometimese NaNs without the fix).  I'll let CI test the fix more thoroughly.
The larger failure is how the macros were used in the Sedov par files.  A lazy
and ugly use of preprocessor conditionals by me led to slow detection of the
bug.  In particular, I was setting vctr to a 3D value in the case that NDIM
wasn't defined or for any value other than 1 or 2.  This is clearly not what is
intended.  One fix is to only set vctr to the 3D value when NDIM is 3 and have
an else condition throw an error.  The solution adopted improves on this as it
only allows N-D problems to use an N-D par file.

The use of preprocessors throughout the repo should be reviewed as part of this
cleaning (Issue #44) to see if my laziness was widespread.