(Issue #45) Updated the error handling scheme for more tests so that they are similar to what was implemented for the Grid/general and Runtime/cpu tests in the previous commit.
The ThreadTeam build is failing in the GCE test suite without GitLab CI
registering the failure.  I have confirmed that the build script fails with a
non-zero exit code.  Changed the GitLab CI script so that the build is executed
on one line; the test, execution on a separate line.  Hopefully, the calling of
both on a single line is the cause of the issue.