(Issue #28) As I have been adding features to the test setup script, I have been adding in more and more modern python (e.g., pathlib).  I decided to go ahead and propagate such changes throughout so that the code isn't a Frankenstein.
This was coupled with updating the script so that it no longer links Makefiles
in, but rather copies them in.  Also, it now removes write permissions for all
users for all files in a newly created build folder.  These changes are
motivated by the idea that a build folder can now be used for as many jobs as
desired, that all results are stored in the build folder (at least on Summit),
and therefore that the contents of the build folder are important metadata to
protect and preserve.

I am testing this as I update all Summit build/job scripts.