(Issue #43) WIP: Updated the build system so that it can be used to build Milhoja libraries that use the CUDA backend with OpenACC offloading.
Moved the CUDA backend headers to dedicated folder so that they are only copied
into a library installation that uses the CUDA backend.

The Runtime/cpu and /gpu tests were built in a somewhat lazy way which used
the presence of the CUDA backend to imply that the tests should use the GPU.
Split up the tests in a cleaner way.  The duplicated tests are few and
hopefully won't change much in the future.

Some manual testing on Summit to confirm only that tests build and run
without error.