(Issue #43) Implemented a minimal Fortran interface for just the Grid functionality.  This was an adaptation and cleaning of the original prototype on the 43cleanup branch.
I was able to successfully build and pass a new Milhoja/Grid unittest in Flash5
using intel on GCE.  With this, it seems that the Fortran/C++ infrastructure is
seeded.  Ideally, the rest of this interface will be developed with TDD using
the Flash5 unittest.  This should allow for more careful and easier development.