(Issue #52) WIP: Prefer Milhoja setup macros defined in Milhoja.h.
Note that the global macros defined by the build system are still in place.  No
effort made to see if the Milhoja tests need to be updated as well.

Ran [23]Sedov problems on Summit with the CPU-only and GPU-only config.  Both
built and ran through with a reasonable number of total steps.  No further
checking done.