Add in first working version of TileWrapper.
This is a rudimentary draft that allows me to use the TileWrapper class in the
Sedov Milhoja/CPU test problem.  The implementation will need cleaning.

Ran the test problem on GCE/c04.  The library and test compiled without new
warnings.  The test ran through well and appropriate IQ were conserved.