Fixed jsons in variant folders, added new class to help with data formatting, updated docs
Added the DataPacketMemberVars class for use with the packet generator. This class serves to pack package different variables easier since the names of each variable are derived from the name of the item in the data packet with a predetermined prefix and suffix. Keeping this information contained in a separate class is better organization than using string constants, and less error prone than having magic strings thrown everywhere in the script and related files. The new generation has not been tested, but the output files for CPP packets have been mostly the same minus a small name change that's consistent across the whole file. This class has also not been integrated with the cpp2c generator, but the c2f generator likely doesn't need this class since the script essentially uses its own naming conventions for the function declaration and implementation. Python development on any unix text editor is a nightmare for me so I'm committing this from the cels note so it can be pulled and tested on summit.