Added a few changes to the generation of data packets to better support the generated flash-x data packet JSONs. Please note: the changes are not specific to flash-x, and just allow more leniency when it comes to information specified in the JSON.
These changes are not flash-x specific and only serve to allow less information to be specified inside of datapacket JSONs. This mostly pertains to the start and end index for tile_in and tile_in_out arrays. Since the data packet only really cares about the sizes (kind of), if start and indices are not specified the array is considered to be size 1. These changes have not been tested with existing data packets, but due to the implementation the new code should not effect existing packets. These changes also allow for boolean / logical variables to be passed into the JSON, something that I completely forgot about when creating the generator. This adds another item to the sizes json, since the size of booleans are implementation specific.