Removed nscratch bytes from computation of nBytesPerPacket to be more precise.
From a conceptual standpoint, the scratch data is not in the data packet. It doesn't get moved to or from the host device or remote device, even though the data packet needs to be aware of it. Thus, the calculation of the total size of the data packet has been altered slightly to not include the size of the scratch data, but the scratch data is included when asking the runtime for memory for the data packet.