Move generation into generic functions in python package.
Built out/simplified command line tools with this change.  The main workhorses
are not generate_data_item and generate_task_function.  These should generate
the correct code based on the contents of the TF spec file.  Added in two
utility functions that might come in handy if we update the TF specs.  These
latter two are untested and fail if run.

Ran runtime/cpu and Sedov/cpu on GCE/c04.  No compiler warnings and all
appeared to work.