Use data_item instead of device in Milhoja-JSON.
This seems to be an improvement for the JSON file as it more explicitly
expresses that we deal with (task function, data_item) pairs.  Also, individual
config values are placed into groups in a more precise way after the change.

We now have a scheme for managing whether or not code files are overwritten by
generators.  Basically, programs and their users decide this now and have some
flexibility for how to do this.

As a result of many of these changes, the code generation scripts for the test
build system now need to be in python so that they can use the data in the JSON
files.

Built and successfully ran Runtime/CPU and Sedov/cpu/2D with GCC.  No new
compiler warnings and all looked good including the generated Makefiles.