Simplify grid_data specification.
The previous version of the grid_data specification had extra information in
it.  For instance, Milhoja owns the Grid/Tile interfaces.  Therefore, it should
be able to determine the extents and lbound information automatically.  Once
the grid information was added to the specification, this become possible.
Made necessary changes here.

Testing passing as expected.