A whole bunch of changes
Transitioning the data packet generator to a class within the milhoja pypackage is incomplete. The goal for now is to create an interface that other python code can use in order to generate data packets in accordance with the other milhoja code generation tools. The mechanisms of the overall generator will be split up into multiple files for now, but I'm okay with that as long as there is a uniform way to call any code generator that a developer needs to call. This code will not be pretty and it doesn't need to be, for now. It's important to me that the milhoja code generation tools are easy to use, and having the data packet generator be separate from the other generators is cumbersome. Since I'm already testing the generators on this branch anyway, I figured I would go ahead and use this branch as a means to move my generator into the milhoja pypackage.