Added initial solutions for collapsing argument specifications and argument mapping
The initial solution to this includes a few assumptions. 1) External and scratch data has some key that returns the name of the shared variable it pulls from. 2) We have full control over the names of the variables that get stored inside of the task function json. This code is also unoptimized, as it moves through each type of variable instead of checking all variables in the node graph before moving to the next.