Replace old jsons with new tf jsons (again?), removed some unused functions, fixed scratch array generation in packet generators, updated tests.
The new packets have not been tested with the simulations. The base test packets are identical to the old packets. Generation for sedov packets has changed slightly now that lbound is processed. Lbound strings needed a slight modification in order to be easily parsable, they now need to always have a set of parenthesis on the outside of the string. Doing that makes them much easier to parse.