Pass variable base index on to TaskFunction class.
We should allow applications to specify variable access patterns using whatever
index set they desire.  We should also allow this to vary from one operation
specification to the next.  Since the TF spec also uses variable indices and we
allow for applications to write these directly if they so choose, we should
treat the variables in the same way.  The TaskFunction class will then have to
convert the indices to the base assumed by the DataPacket generator so that it
only deals with one base.