Removed generate_helpers_tpl and cpp_helpers files in favor of TemplateUtility classes
These utility classes provide an interface for helping create templates for use with cgkit. This should make it easier to extend or change functionality for specific languages, and also simplifies the logic contained inside of each iterate function at the cost of some duplicate code. Currently, there is only a template utility class for Cpp packets, but since the classes from existing code it should be easy to create the fortran one. The tests that I have written have passed for all Runtime tests, Sedov tests have not been tested.