Added SubroutineGroup as a class that replaces operation spec concept.
Notion of subroutine group more general than operation, which is likely
Flash-X-specific.  The class was carefully designed to be read-only with simple
interface.  Use of this has simplified the TaskFunctionAssembler some and its
code is not so difficult since it can no longer alter the original specs.

This is under heavy testing, which all pass, and has 100% coverage by line.