Grid specification removed from subroutine group specification.
It didn't make sense to include that in the groups.  It was especially
problematic because it should be true that all group specifications would use
the same grid specification.  It's passed in separately now but still at the
same time as the group specs as these should be established by the application
at the same time.  Also, it could be that the grid specification is needed to
test some of the group specifications.

All package tests passing.  Ran some Grid, Runtime/CPU, and Sedov/CPU tests on
c04.  No obvious failures.