Start integrating Sedov/gpu/variant3 in 2D and 3D.
The driver and TF had been simplified to just 2D.  I therefore reintroducted 2D
but without simultaneous launching of the compute flux routines.  The 3D JSON
for the problem was likely just some version that I had created to test out the
JSON specification format.  I updated for C++.

Ran 2D on Summit.  It ran 1000 steps and has reasonable conservation of the
necessary physical quantities.  Log file shows that hydro was being run on CPU
and GPU in parallel.  The 3D problem does not compile yet.  This makes sense
since we likely have never tested automatic packet generation in 3D for this
problem.  It should be assumed that the issue might be with the JSON spec that
I updated.