Get sizes.json from library installation.
This appears to be working.  However, the setup.py script presently uses the
same file in the repo regardless of platform.  This can easily be updated once
the file is installed by configure.py for each Milhoja installation.

Manual testing on summit showed correct debug output.  I was able to build and
run all tests except for Grid/general.  I suspect that that failure is due to
an unmaintained build script.