Finished integrating sizes.json into installation & test builds.
Reverted change to GridAmrex.  I suspect that someone is using an old AMReX
library somewhere.  I need to discuss with Wesley to identify where those
libraries are.  As part of this, I made the library Makefile more robust by
basically rebuilding everything if any Milhoja header file is altered.  This is
certainly a conservative approach, but seems acceptable since build times are
not long.

Tested this by updating file timestamps for the following files with "touch"
and then rerunning make to see that the sizes.json file is regenerated with a
rebuilt program:
* Summit's site Makefile
* Main library Makefile
* ./tools/createSizesJson.cpp
* ./build/Milhoja.h
* ./includes/Milhoja_real.h
Rebuilt Milhoja libraries and confirmed that all installs have a sizes.json
where expected.  Built and successfully ran Runtime/GPU on Summit.