Updated cpp2c generated, added improved lbound parsing function that handles some more cases.
I added this new lbound parser function because I noticed that the cpp2c layer for a tile wrapper w/ a fortran tf required a more complex lbound parsing function. This new function is simpler and handles most of the same cases that the original does, but it does a few things differently. First, it returns a list of strings of each index in the bound. This means that the metadata strings included in the lbound string will be split into their I,J, and K parts. This allows for more flexibility when parsing, and the generators that used the old parser only need to combine each part into its own intvect. In this case, the parsing of the lbound and the function have simpler implementations, so some of the complexity is handed off to the actual generators. There's also a small issue of having to replace any variable names inside of the lbound string with their appropriate source value. The generators will need to replace the source values with the associated variable name once the lbound list is returned. There isn't much of a way around this, unless the calling code passes a mapping of name to source to the function.