Allow tile_interior/tile_arrayBounds to be TF arguments.
For now, the best decisions for dealing with this metadata is to include them
in data packets.  In particular, we will allow for now the possibility that a
data packet contains lo, hi, and interiors.  Once we have finished all code
generators we can revisit this issue and determine if there is a way to
optimize the data packets for such situations.