Tweaks re level convention and scratch-free Fortran routines
A refinement `level` integer should be interpreted
 * as 0-based within the Milhoja library code,
 * as 1-based in the general Flash-X code.
So convert when Milhoja constructs a Milhoja::TileFlashxr object from
tileCInfo components.

It is completely okay for a static Fortran routine (that is called
from an ORCHA task function) to not have any arguments of source type
scratch. Remove code that blocks this from working.