Define JJ(base) and KK(base) methods in milhoja::IntVect
These are alternatives of J() and K(), respectively.
Where J and K return 0 for inactive directions,
JJ(base) and KK(base) return the value of base instead.
It is expected that base will usually be 1 when JJ and KK
are used in code for Fortran, and base will be 0 if JJ and KK
are used in code for C or C++ task functions.