Correct indexes beyond the first three ones
The logic that fills in index values for indexes beyond the first three
for source=lbound arguments now sets them to 1, rather than
to some variable value.

Maybe this should be done differently in case code for interfacing to
a language other than Fortrasn is generated?  However, (1) the
"constant" pattern GRID_DATA_LBOUNDS that is changed here appears to
be used only on code for Fortran now, and (2) the previous version of
the code appears to have been wrong even for a non-Fortran use (as
long as such a non-Fortran use would be aalong the same lines as the
existing Fortran use).

This addresses Item 4 of OrchestrationRuntime Issue #94.