Use JJ(1) and KK(1) where needed for Fortran code
Use them, instead of J() and K() in the previous code, where
such a change appears necessary to generated proper index
values for inactive directions - at least for Fortran subroutines
called from Fortran task functions.

There could be cleaner, or otherwise better, ways to do this.

This patch goes to some trouble to leave things alone (i.e.,
index values for inactive directions continue to be returned
as 0) where a change is not known to be necessary for Fortran,
in order not to disturb tests and applications cases that do
**not** involve Fortran.

Changes as in this patch are needed, because without them it was
found that index values in lo, hi, etc., ended up being inconsistent
with the actual effective bounds of array passed in to Fortran
routines as argument. They **must** be consistent, otherwise
code like the following does not work in those Fortran routines
(in particular for 2D and 3D cases):

   do k = lo(KAXIS), hi(KAXIS)
      do j = lo(JAXIS), hi(JAXIS)
         do i = lo(IAXIS), hi(IAXIS)
           ... Do something with solnData(i,j,k,var) ...

Inactive arrays values must be set to 1 for Fortran code
(i.e., k ranges from 1 to 1 in the above loop nest, not
from 0 to 0), because that is the convention used throughout
the Flash-X code.

This addresses Item 1 of OrchestrationRuntime Issue #94.