use subroutine wrappers to avoid array slices
Profiling results from the Nsight Systems reports that
a huge amount (~90%) of timing is consumed by
`cuMemHostAlloc`, when a slice of array (e.g., `Uin(:, :, :, :, n)`)
is passed to the SFR's argument.

This commit avoids `cuMemHostAlloc` in the profiling results
by introducing "wrapper" subroutines for each SFRs.
However, the overall performance results remain the same,
even though the Nsight Systems doesn't report `cuMemHostAlloc`.
Perhaps it was an incorrect profiling results, but I push this commit
to investigate this further.