Restore main branch to before the last 4 commits.
Those cherry-pick commits should not have been done on the main branch.
The shell command I am using to restore the main branch to the desired state is

  git restore --staged --worktree --source=1f5a60a88315ca854516b254d5796b8e81d8d3f6 :/

A proper merge commit, merging the GitHub PR #1, is to follow soon.
The changes undone here will then be applied once more, together with others.