Tweak the logic of creating POINTERs to pass for some arguments
Fixes for some cases, avoiding either wrong code (including
compilation errors) or applying the "intermediate pointer" mechanism
in situations where it is not needed.

NOTE that the recognition of some subroutine as ones that have POINTER
array dummy arguments is currently hardwired, as recognizing only
subroutines that have either "Eos_guardCells" or "Eos_wrapped" in the
name! There should be a better way, it should not be coupled to a
specific state of the Flash-X code!