Recognizing subroutine arguments marked as 'source=verbatim'
In the annotation for Milhoja in a Fortran interface module source,
a new source type verbatim is now supported.

This can be used to pass simple scalar constants directly to a
Fortran subroutine when it is called from a Milhoja task function.
Previously, source=external arguments have been used for achieving
this, but that i unnecessarily complicated for simple cases, since
it requires adding additional variables (including, internally,
class members of data items in the Milhoja library  implementation)
at several levels, just for passing a value into a task function
and onward to the Fortran subroutine.

The specification of the argument value currently has to look like
this:

    value=literal:123

The special string "literal:" is currently necessary.
Support for kinds of value other than "literal" could be added
later if useful.

It turns out that this also can be used for purposes more complex
that simple scalar constants. Expressions and array values are
possible. (The syntax for array constants here currently requires
delimiters '(/' and '/)' rather than '[' and ']' or '(' and ')'.)
If the writer of interface annotations uses these complicated forms,
or variables or expressions thar are not constants, no checking
of validity is done; proceed at your own risk.

Note that Fortran/C or Flash-X preprocessor macros can be used,
as for other ORCHA interface argument annotations given
in a file that gets preprocessed.

Some example uses, in the form that can occur in interface annotations:

   !!   _stage   :: source=verbatim, value=literal:1
   !!   _eosMode :: source=verbatim, value=literal:MODE_DENS_EI
   !!   _loLocal :: source=verbatim, value=literal:(/GRID_ILO, GRID_JLO, GRID_KLO/)
   !!   _hiLocal :: source=verbatim, value=literal:tile_hi(1:MDIM)-tile_lbound(1:MDIM)+1

(Not all of these are claimed to be useful or recommended, but they
have been somewhat tested.)